Move ssr preview to react preset
