+++ package.json
@@ -40,8 +40,12 @@
     "date-fns": "^2.14.0",
     "eslint": "^6.6.0",
     "eslint-config-nebenan": "^1.3.2",
+    "eval": "^0.1.4",
+    "express": "^4.17.1",
+    "is-object": "^1.0.1",
     "lerna": "^3.20.2",
     "lodash": "^4.17.15",
+    "mini-css-extract-plugin": "^0.9.0",
     "mocha": "^8.0.1",
     "nebenan-helpers": "^4.2.0",
     "nebenan-ui-kit": "^4.12.0",
@@ -59,7 +63,12 @@
     "sass": "^1.26.8",
     "sass-functions": "^1.4.0",
     "sass-loader": "^8.0.2",
-    "style-loader": "^1.2.1"
+    "style-loader": "^1.2.1",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-middleware": "^3.7.2",
+    "webpack-dev-server": "^3.11.0",
+    "webpack-hot-middleware": "^2.25.0"
   },
   "scripts": {
     "lint": "lerna run lint",
