+++ presets/react/storybook/ssr/src/utils.jsx
@@ -1,7 +1,7 @@
 import qs from 'qs';
 
 export const getStories = () => {
-  const stories = require.context('../../../src', true, /\.stories\.jsx$/);
+  const stories = require.context(`${ROOT_PATH}/src`, true, /\.stories\.jsx$/);
   return stories.keys().map(stories);
 };
 
@@ -9,10 +9,10 @@ export const getStories = () => {
 export const getUrlForStory = ({ storyId, viewMode }) => {
   const { path, selectedKind, selectedStory, ...rest } = qs.parse(document.location.search, {
     ignoreQueryPrefix: true,
-  })
+  });
   return `${document.location.pathname}?${qs.stringify({
     ...rest,
     id: storyId,
     viewMode,
-  })}`
-}
+  })}`;
+};
